pyproject.toml
[project]
name = "context-forge-demo"
version = "0.1.0"
description = "Demo of ContextForge MCP Gateway with entreprise use case"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "fastmcp>=2.14.5",
    "json-repair>=0.57.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ibm>=1.0.4",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.9",
    "langfuse>=3.14.1",
    "langgraph>=1.0.8",
    "langgraph-cli[inmem]>=0.4.12",
    "pre-commit>=4.5.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
ingest = [
    "docling>=2.73.1",
    "langchain-docling>=2.0.0",
]
wxo = [
    "ibm-watsonx-orchestrate>=2.6.0",
]
